requirements.txt
requests
dnspython
tldextract
colorama
beautifulsoup4